fix admin && UI customer

+++ Library/src/main/java/com/ecommerce/library/dto/ProductDto.java
@@ -5,18 +5,30 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+import java.text.DecimalFormat;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDto {
     private Long id;
+
     private String name;
+
     private String description;
-    private int costPrice;
-    private int salePrice;
+
+    private double costPrice;
+
+    private double salePrice;
+
     private int currentQuantity;
+
     private Category category;
+
     private String image;
+
     private boolean active;
+
     private boolean deleted;
 }
